+++ src/Http/Http.Abstractions/src/PublicAPI.Unshipped.txt
@@ -38,26 +38,22 @@ Microsoft.AspNetCore.Http.IValueHttpResult
 Microsoft.AspNetCore.Http.IValueHttpResult.Value.get -> object?
 Microsoft.AspNetCore.Http.IValueHttpResult<TValue>
 Microsoft.AspNetCore.Http.IValueHttpResult<TValue>.Value.get -> TValue?
+Microsoft.AspNetCore.Http.IProblemDetailsService
+Microsoft.AspNetCore.Http.IProblemDetailsService.WriteAsync(Microsoft.AspNetCore.Http.ProblemDetailsContext! context) -> System.Threading.Tasks.ValueTask
+Microsoft.AspNetCore.Http.IProblemDetailsWriter
+Microsoft.AspNetCore.Http.IProblemDetailsWriter.TryWriteAsync(Microsoft.AspNetCore.Http.ProblemDetailsContext! context) -> System.Threading.Tasks.ValueTask<bool>
 Microsoft.AspNetCore.Http.Metadata.IFromFormMetadata
 Microsoft.AspNetCore.Http.Metadata.IFromFormMetadata.Name.get -> string?
-Microsoft.AspNetCore.Http.Metadata.IProblemDetailsMetadata
-Microsoft.AspNetCore.Http.Metadata.IProblemDetailsMetadata.ProblemType.get -> Microsoft.AspNetCore.Http.ProblemDetailsTypes
-Microsoft.AspNetCore.Http.Metadata.IProblemDetailsMetadata.StatusCode.get -> int?
 Microsoft.AspNetCore.Http.Metadata.IRequestSizeLimitMetadata
 Microsoft.AspNetCore.Http.Metadata.IRequestSizeLimitMetadata.MaxRequestBodySize.get -> long?
 Microsoft.AspNetCore.Http.ProblemDetailsContext
 Microsoft.AspNetCore.Http.ProblemDetailsContext.AdditionalMetadata.get -> Microsoft.AspNetCore.Http.EndpointMetadataCollection?
 Microsoft.AspNetCore.Http.ProblemDetailsContext.AdditionalMetadata.init -> void
 Microsoft.AspNetCore.Http.ProblemDetailsContext.HttpContext.get -> Microsoft.AspNetCore.Http.HttpContext!
+Microsoft.AspNetCore.Http.ProblemDetailsContext.HttpContext.init -> void
 Microsoft.AspNetCore.Http.ProblemDetailsContext.ProblemDetails.get -> Microsoft.AspNetCore.Mvc.ProblemDetails!
 Microsoft.AspNetCore.Http.ProblemDetailsContext.ProblemDetails.init -> void
-Microsoft.AspNetCore.Http.ProblemDetailsContext.ProblemDetailsContext(Microsoft.AspNetCore.Http.HttpContext! httpContext) -> void
-Microsoft.AspNetCore.Http.ProblemDetailsTypes
-Microsoft.AspNetCore.Http.ProblemDetailsTypes.All = Microsoft.AspNetCore.Http.ProblemDetailsTypes.Server | Microsoft.AspNetCore.Http.ProblemDetailsTypes.Routing | Microsoft.AspNetCore.Http.ProblemDetailsTypes.Client -> Microsoft.AspNetCore.Http.ProblemDetailsTypes
-Microsoft.AspNetCore.Http.ProblemDetailsTypes.Client = 4 -> Microsoft.AspNetCore.Http.ProblemDetailsTypes
-Microsoft.AspNetCore.Http.ProblemDetailsTypes.None = 0 -> Microsoft.AspNetCore.Http.ProblemDetailsTypes
-Microsoft.AspNetCore.Http.ProblemDetailsTypes.Routing = 2 -> Microsoft.AspNetCore.Http.ProblemDetailsTypes
-Microsoft.AspNetCore.Http.ProblemDetailsTypes.Server = 1 -> Microsoft.AspNetCore.Http.ProblemDetailsTypes
+Microsoft.AspNetCore.Http.ProblemDetailsContext.ProblemDetailsContext() -> void
 Microsoft.AspNetCore.Routing.RouteValueDictionary.RouteValueDictionary(Microsoft.AspNetCore.Routing.RouteValueDictionary? dictionary) -> void
 Microsoft.AspNetCore.Routing.RouteValueDictionary.RouteValueDictionary(System.Collections.Generic.IEnumerable<System.Collections.Generic.KeyValuePair<string!, object?>>? values) -> void
 Microsoft.AspNetCore.Routing.RouteValueDictionary.RouteValueDictionary(System.Collections.Generic.IEnumerable<System.Collections.Generic.KeyValuePair<string!, string?>>? values) -> void
