--- src/Http/ProblemDetails/src/PublicAPI.Shipped.txt
@@ -1,1 +1,0 @@
-#nullable enable
