--- src/Http/ProblemDetails/src/PublicAPI.Unshipped.txt
@@ -1,10 +1,0 @@
-#nullable enable
-Microsoft.AspNetCore.Http.ProblemDetailsOptions
-Microsoft.AspNetCore.Http.ProblemDetailsOptions.AllowedProblemTypes.get -> Microsoft.AspNetCore.Http.ProblemDetailsTypes
-Microsoft.AspNetCore.Http.ProblemDetailsOptions.AllowedProblemTypes.set -> void
-Microsoft.AspNetCore.Http.ProblemDetailsOptions.ConfigureDetails.get -> System.Action<Microsoft.AspNetCore.Http.HttpContext!, Microsoft.AspNetCore.Mvc.ProblemDetails!>?
-Microsoft.AspNetCore.Http.ProblemDetailsOptions.ConfigureDetails.set -> void
-Microsoft.AspNetCore.Http.ProblemDetailsOptions.ProblemDetailsOptions() -> void
-Microsoft.Extensions.DependencyInjection.ProblemDetailsServiceCollectionExtensions
-static Microsoft.Extensions.DependencyInjection.ProblemDetailsServiceCollectionExtensions.AddProblemDetails(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services) -> Microsoft.Extensions.DependencyInjection.IServiceCollection!
-static Microsoft.Extensions.DependencyInjection.ProblemDetailsServiceCollectionExtensions.AddProblemDetails(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, System.Action<Microsoft.AspNetCore.Http.ProblemDetailsOptions!>! configureOptions) -> Microsoft.Extensions.DependencyInjection.IServiceCollection!
