--- src/Http/ProblemDetails/test/Microsoft.AspNetCore.Http.ProblemDetails.Tests.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>$(DefaultNetCoreTargetFramework)</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="Microsoft.AspNetCore.Http.Extensions" />
-    <Reference Include="Microsoft.AspNetCore.Http.ProblemDetails" />
-  </ItemGroup>
-
-</Project>
