+++ src/Http/Routing/src/Matching/HttpMethodMatcherPolicy.cs
@@ -4,7 +4,6 @@
 using System.Linq;
 using System.Runtime.InteropServices;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Metadata;
 using Microsoft.Extensions.Primitives;
 using Microsoft.Net.Http.Headers;
 
@@ -16,8 +15,6 @@ namespace Microsoft.AspNetCore.Routing.Matching;
 /// </summary>
 public sealed class HttpMethodMatcherPolicy : MatcherPolicy, IEndpointComparerPolicy, INodeBuilderPolicy, IEndpointSelectorPolicy
 {
-    private static readonly EndpointMetadataCollection Http405EndpointMetadata = new(new RoutingProblemMetadata(StatusCodes.Status405MethodNotAllowed));
-
     // Used in tests
     internal static readonly string PreflightHttpMethod = HttpMethods.Options;
 
@@ -413,7 +410,7 @@ public sealed class HttpMethodMatcherPolicy : MatcherPolicy, IEndpointComparerPo
                 context.Response.StatusCode = StatusCodes.Status405MethodNotAllowed;
                 return Task.CompletedTask;
             },
-            Http405EndpointMetadata,
+            EndpointMetadataCollection.Empty,
             Http405EndpointDisplayName);
     }
