+++ src/Http/Routing/src/Microsoft.AspNetCore.Routing.csproj
@@ -32,7 +32,6 @@
     <Compile Include="$(SharedSourceRoot)MediaType\HttpTokenParsingRule.cs" LinkBase="Shared" />
     <Compile Include="$(SharedSourceRoot)ApiExplorerTypes\*.cs" LinkBase="Shared" />
     <Compile Include="$(SharedSourceRoot)RoutingMetadata\AcceptsMetadata.cs" LinkBase="Shared" />
-    <Compile Include="$(SharedSourceRoot)ProblemDetails\RoutingProblemMetadata.cs" LinkBase="Shared" />
   </ItemGroup>
 
   <ItemGroup>
