+++ src/Middleware/Diagnostics/test/FunctionalTests/StatusCodeSampleTest.cs
@@ -6,6 +6,8 @@ using System.Net.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Net.Http.Json;
 using Microsoft.AspNetCore.WebUtilities;
+using Microsoft.Net.Http.Headers;
+using System.Net.Http.Headers;
 
 namespace Microsoft.AspNetCore.Diagnostics.FunctionalTests;
 
@@ -23,6 +25,7 @@ public class StatusCodeSampleTest : IClassFixture<TestFixture<StatusCodePagesSam
     {
         // Arrange
         var request = new HttpRequestMessage(HttpMethod.Get, "http://localhost/errors/417");
+        request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("text/plain"));
 
         // Act
         var response = await Client.SendAsync(request);
@@ -39,6 +42,7 @@ public class StatusCodeSampleTest : IClassFixture<TestFixture<StatusCodePagesSam
         //Arrange
         var httpStatusCode = 541;
         var request = new HttpRequestMessage(HttpMethod.Get, $"http://localhost/?statuscode={httpStatusCode}");
+        request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("text/plain"));
 
         //Act
         var response = await Client.SendAsync(request);
@@ -58,6 +62,7 @@ public class StatusCodeSampleTest : IClassFixture<TestFixture<StatusCodePagesSam
         //Arrange
         var httpStatusCode = 400;
         var request = new HttpRequestMessage(HttpMethod.Get, $"http://localhost/?statuscode={httpStatusCode}");
+        request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("text/plain"));
 
         //Act
         var response = await Client.SendAsync(request);
@@ -78,7 +83,8 @@ public class StatusCodeSampleTest : IClassFixture<TestFixture<StatusCodePagesSam
     {
         // Arrange
         var httpStatusCode = 400;
-        var request = new HttpRequestMessage(HttpMethod.Get, $"http://localhost/status?includeProblemMetadata=true&statuscode={httpStatusCode}");
+        var request = new HttpRequestMessage(HttpMethod.Get, $"http://localhost?statuscode={httpStatusCode}");
+        request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
 
         // Act
         var response = await Client.SendAsync(request);
