+++ src/Mvc/Mvc.Core/src/Infrastructure/DefaultProblemDetailsFactory.cs
@@ -13,14 +13,14 @@ namespace Microsoft.AspNetCore.Mvc.Infrastructure;
 internal sealed class DefaultProblemDetailsFactory : ProblemDetailsFactory
 {
     private readonly ApiBehaviorOptions _options;
-    private readonly Action<HttpContext, ProblemDetails>? _configure;
+    private readonly Action<ProblemDetailsContext>? _configure;
 
     public DefaultProblemDetailsFactory(
         IOptions<ApiBehaviorOptions> options,
         IOptions<ProblemDetailsOptions>? problemDetailsOptions = null)
     {
         _options = options?.Value ?? throw new ArgumentNullException(nameof(options));
-        _configure = problemDetailsOptions?.Value?.ConfigureDetails;
+        _configure = problemDetailsOptions?.Value?.CustomizeProblemDetails;
     }
 
     public override ProblemDetails CreateProblemDetails(
@@ -98,6 +98,6 @@ internal sealed class DefaultProblemDetailsFactory : ProblemDetailsFactory
             problemDetails.Extensions["traceId"] = traceId;
         }
 
-        _configure?.Invoke(httpContext!, problemDetails);
+        _configure?.Invoke(new() { HttpContext = httpContext!, ProblemDetails = problemDetails });
     }
 }
